+++ tests/test_get_triangles.py
@@ -18,7 +18,7 @@ class TriangleTester(TestBase):
         self.assertTrue(verts.shape[1] == 3)
         self.assertTrue(verts.dtype == np.float32)
         self.assertTrue(tris.shape[1] == 3)
-        self.assertTrue(tris.dtype == np.int)
+        self.assertTrue(tris.dtype == np.int32)
         self.assertTrue(np.all(tris < len(verts)))
 
     def run_test_with_pathname(self, file, solid):
